requirements.txt
Pillow==10.1.0
Pillow==10.1.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.202
